created delete_address endpoint, deleted adresses_router.py and transfered the get_address to the addresses_router.py, changed the database name in .env to nrep_database back (it was nrep_datadase2)
