+++ .env
@@ -7,5 +7,5 @@ DATABASE_USERNAME=postgres
 DATABASE_PASSWORD=1234
 # DATABASE_HOST=nrep-database.clm4aqceyhvz.us-east-1.rds.amazonaws.com
 DATABASE_HOST=localhost
-DATABASE_NAME=nrep_datadase2
+DATABASE_NAME=nrep_database
 DATABASE_PORT=5432
