+++ app/sql_queries/address/delete_address.sql
@@ -1,0 +1,2 @@
+UPDATE users SET address_id = NULL WHERE address_id = :address_id;
+DELETE FROM addresses WHERE address_id = :address_id;
